+++ enclone_visual/src/gui.rs
@@ -499,7 +499,7 @@ impl Application for EncloneVisual {
             .height(Length::Units(100))
             .scrollbar_width(12)
             .scroller_width(12)
-            .style(style::Squeak)
+            .style(style::ScrollableStyle)
             .push(Text::new(&self.output_value).font(DEJAVU_BOLD).size(13));
 
         // Fix the height of the SVG.  This needs to be set so that there is enough room for
@@ -753,70 +753,4 @@ async fn compute() -> Result<(), String> {
 async fn flash_copy_image_button() -> Result<(), String> {
     thread::sleep(Duration::from_millis(400));
     Ok(())
-}
-
-// ▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓
-
-mod style {
-
-    pub struct RuleStyle;
-
-    impl iced::rule::StyleSheet for RuleStyle {
-        fn style(&self) -> iced::rule::Style {
-            iced::rule::Style {
-                color: Color::from_rgb(0.0, 1.0, 1.0),
-                width: 3,
-                radius: 1.0,
-                fill_mode: iced::rule::FillMode::Percent(100.0),
-            }
-        }
-    }
-
-    pub struct Squeak;
-
-    use iced::{scrollable, Color};
-
-    impl scrollable::StyleSheet for Squeak {
-        fn active(&self) -> scrollable::Scrollbar {
-            scrollable::Scrollbar {
-                background: Color::from_rgb(0.75, 0.75, 0.75).into(),
-                border_radius: 2.0,
-                border_width: 0.0,
-                border_color: Color::TRANSPARENT,
-                scroller: scrollable::Scroller {
-                    color: Color::from_rgb(0.0, 0.0, 0.0),
-                    border_radius: 2.0,
-                    border_width: 0.0,
-                    border_color: Color::TRANSPARENT,
-                },
-            }
-        }
-
-        fn hovered(&self) -> scrollable::Scrollbar {
-            let active = self.active();
-            scrollable::Scrollbar {
-                background: Color {
-                    a: 0.5,
-                    ..Color::from_rgb(0.0, 0.0, 0.0)
-                }
-                .into(),
-                scroller: scrollable::Scroller {
-                    color: Color::from_rgb(0.0, 0.0, 0.0),
-                    ..active.scroller
-                },
-                ..active
-            }
-        }
-
-        fn dragging(&self) -> scrollable::Scrollbar {
-            let hovered = self.hovered();
-            scrollable::Scrollbar {
-                scroller: scrollable::Scroller {
-                    color: Color::from_rgb(0.0, 0.0, 0.0),
-                    ..hovered.scroller
-                },
-                ..hovered
-            }
-        }
-    }
 }
