+++ enclone_visual/src/lib.rs
@@ -22,6 +22,7 @@ pub mod enclone_client;
 pub mod enclone_server;
 pub mod geometry;
 pub mod gui;
+pub mod style;
 pub mod svg_to_geometry;
 pub mod update_restart;
